Tarefa2_2022li1g124

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

estendeMapa

A função estendeMapa irá analisar o Mapa que lhe foi submetido e, com o valor Inteiro recebido (dentro do intervalo [1,100]) irá devolver o mesmo Mapa recebido, mas adicionará uma "linha" "pseudo-aleatoriamente" usando o Inteiro recebido como recurso (irá devolver o mesmo Mapa, mas adicionando um par de Terreno e lista de Obstáculos)

Podemos assim definir estendeMapa da seguinte maneira:

Definição dada

estendeMapa :: Mapa -> Int -> Mapa
estendeMapa (Mapa l []) x = Mapa l [(Relva, primlin l [])]  
estendeMapa (Mapa l terobs) x 
      = let proxTer = proximoTerrenoVel x (Mapa l terobs)
            proxObs = proximoObstaculo x l ((proxTer, []):(terobs))
        in Mapa l (((proxTer), (proxObs)):terobs)

    :: Mapa  assume-se qualquer Mapa a esta função
    -> Int   iremos tambem providenciar-lhe um número qualquer
                 que será usado para escolher um Terreno qualquer
                 e uma lista de Obstáculos qualquer (tudo aleatóriamente)
    -> Mapa  irá devolver um mesmo Mapa, com uma "linha adicionada"
                 (adicionando um par Terreno | Lista de Obstáculos à cabeça
                 da lista de pares)

Caso o Mapa dado tenha um conjunto vazio de Terrenos | Lista de Obstáculos, a função irá devolver um Terreno "Relva" com uma lista de Obstáculos preenchida de Obstáculos "Nenhum", baseando-me no Jogo original (iremos evitar que o Jogador comece num "Rio " na mesma coordenada de um "Nenhum", ou numa "Estrada" na mesma coordenada de um "Carro"). Usando uma função auxiliar "primlin"

Caso contrário, irá criar uma nova "linha" com auxílio das funções "proximoTerrenoVel" (para determinar o próximo Terreno) e "proximoObstaculo" (para determinar uma nova lista de Obstáculos)

Exemplo (para um Mapa sem Terrenos ou Obstáculos)

>>> estendeMapa (Mapa 6 []) 10
Mapa 6 [(Relva,[Nenhum,Nenhum,Nenhum,Nenhum,Nenhum,Nenhum])]

primlin :: Int -> [Obstaculo] -> [Obstaculo]

primlin

A função primlin é uma função "cega" por assim dizer. primlin aceita apenas a Largura de um Mapa e uma lista de obstáculos "qualquer", que, recursivamente, será preenchida por Obstáculos "Nenhum" até ter um comprimento (length) igual à Largura do Mapa

Podemos assim definir primlin da seguinte maneira:

Definição dada

primlin :: Int -> [Obstaculo] -> [Obstaculo]
primlin l t | length t == l = t
            | otherwise     = primlin l ((Nenhum):t)

    :: Int          recebe assim a Largura do Mapa
    -> [Obstaculo]  irá receber uma lista de Obstáculos (vinda da função estendeMapa
                        que apenas lhe dá uma lista vazia para, propositadamente, só ser
                        capaz de criar uma lista de obstáculos "Nenhum" recursivamente)
    -> [Obstaculo]  irá devolver uma lista de Obstáculos (respetivamente, uma lista preenchida
                        de Obstáculos "Nenhum"),

proximoTerrenoVel :: Int -> Mapa -> Terreno

proximoTerrenoVel

A função "proximoTerrenoVel" é uma função que, auxiliada pelas funções "randomizer" e "proximosTerrenosValidos", determina qual será o próximo Terreno de uma lista de possíveis Terrenos.

Através da função "randomizer", dependendo do número resultante ([0,1,2]), "proximoTerrenoVel" irá escolher a cabeça, o último membro ou até a cabeça da tail da lista de possíveis Terrenos.

Tal lista de possíveis Terrenos é dada pela função auxiliar "proximosTerrenosValidos"

Podemos assim definir "proximoTerrenoVel" da seguinte maneira:

Definição dada

proximoTerrenoVel :: Int -> Mapa -> Terreno
proximoTerrenoVel x (Mapa l terobs) 
      | random1 == 2 
            = head (terrenosVels x (terreno))
      | random1 == 1 
            = head (tail (terrenosVels x (terreno)))
      | random1 == 0 
            = last (terrenosVels x (terreno))
            where terreno = proximosTerrenosPossiveis (Mapa l (reverse terobs)) 0 0 0
                  random1 = (mod (head(randomizer x)) 3)

    :: Int     recebe assim o número dado pelo Usuário (o dado aleatório)
    -> Mapa    recebe também o Mapa do Jogo presente para a "proximosTerrenosValidos"
                   conseguir perceber quais os Proximos Terrenos Validos
    -> Terreno irá devolver um Terreno possível decidido pela função "randomizer"
                   e escolhido de uma lista providenciada pela "proximosTerrenosValidos"

terrenosVels :: Int -> [Terreno] -> [Terreno]

terrenosVels

A função "terrenosVels" é uma função que, dado um Int e uma lista de Terrenos devolve uma lista de Terrenos possíveis com auxílio da função "proximosTerrenosPossiveis", e com Velocidades que variam do (-3) ao 3

Podemos assim definir "terrenosVels" da seguinte maneira:

Definição dada

terrenosVels :: Int -> [Terreno] -> [Terreno]
terrenosVels _ [] = []
terrenosVels x (Relva:t) = (Relva):(terrenosVels x t)
terrenosVels x ((Estrada k):t) | k == 0    = (Estrada ((posneg x) * (velrandom x))):(terrenosVels x t)
                               | otherwise = (Estrada ((k) * (velrandom x))):(terrenosVels x t)
terrenosVels x ((Rio k):t) | k == 0    = (Rio ((posneg x) * (velrandom x))):(terrenosVels x t)
                           | otherwise = (Rio ((k) * (velrandom x))):(terrenosVels x t)
velrandom x = (mod (head(randomizer (x+207))) 3) + 1

    :: Int      recebe um Mapa para percebe quais os próximos Terrenos possíveis 
    -> [Terreno] irá receber uma lista de Terrenos com Velocidades que variam do
                     (-1) ao 1,
    -> [Terreno] irá devolver uma lista de Terrenos possíveis decidido pela 
                     função "proximosTerrenosPossiveis",

velrandom :: Int -> Int

proximosTerrenosPossiveis :: Mapa -> Int -> Int -> Int -> [Terreno]

proximosTerrenosPossiveis

A função "proximosTerrenosPossiveis" é uma função que, dada um Mapa, e três Inteiros (denominadamente, três 0's) devolve uma lista de Terrenos

Podemos assim definir "proximosTerrenosPossiveis" da seguinte maneira:

Definição dada

proximosTerrenosPossiveis :: Mapa -> Int -> Int -> Int -> [Terreno]
proximosTerrenosPossiveis (Mapa larg (((Rio vel1), obs):[])) _ _ rio | (rio < 3) 
                                                                        = [Rio ((-1)*(signum vel1)), Estrada ((-1)*(signum vel1)), Relva]
                                                                     | otherwise               
                                                                        = [Estrada ((-1)*(signum vel1)), Relva]

proximosTerrenosPossiveis (Mapa larg (((Estrada vel1), obs):[])) estr _ _ | (estr < 3) 
                                                                              = [Rio ((-1)*(signum vel1)), Estrada ((-1)*(signum vel1)), Relva]
                                                                          | otherwise               
                                                                              = [Rio ((-1)*(signum vel1)), Relva]

proximosTerrenosPossiveis (Mapa larg []) _ relv _ | (relv > 4)
                                                      = [Rio 0, Estrada 0]                      
                                                  | otherwise 
                                                      = [Rio 0, Estrada 0, Relva]

proximosTerrenosPossiveis (Mapa larg ((Estrada vel1, obs):t)) estr relv rio 
      = proximosTerrenosPossiveis (Mapa larg t) (estr+1) 0 0

proximosTerrenosPossiveis (Mapa larg ((Relva, obs):t)) estr relv rio       
      = proximosTerrenosPossiveis (Mapa larg t) 0 (relv+1) 0

proximosTerrenosPossiveis (Mapa larg ((Rio vel1, obs):t)) estr relv rio     
      = proximosTerrenosPossiveis (Mapa larg t) 0 0 (rio+1)

    :: Mapa      recebe um Mapa para percebe quais os próximos Terrenos possíveis 
    -> Int       recebe um Int relacionado há quantidade de "Estradas" seguidas na lista
    -> Int       recebe um Int relacionado há quantidade de "Relvas" seguidas na lista
    -> Int       recebe um Int relacionado há quantidade de "'Rios" seguidas na lista
    -> [Terreno] devolve uma lista de Terrenos possíveis 

Baseando-se nas condições propostas na 1ª tarefa do projeto, a função "proximosTerrenosPossiveis" irá recursivamente ver todos os "Terrenos" de todos os pares da lista e irá concluir quais os Terrenos possíveis para continuar o Mapa (tendo em especial atenção a 2ª e 7ª condição da tarefa 1)

proximoObstaculo :: Int -> Int -> [(Terreno, [Obstaculo])] -> [Obstaculo]

proximoObstaculo

A função "proximoObstaculo" é uma função que, dada um Mapa, e três Inteiros (denominadamente, três 0's) devolve uma lista de Terrenos

Podemos assim definir "proximoObstaculo" da seguinte maneira:

Definição dada

proximoObstaculo :: Int -> Int -> [(Terreno, [Obstaculo])] -> [Obstaculo]
proximoObstaculo x l ((ter, obs):t) 
      | elem Tronco obs && length obs == l               = bomRio obs
      | ter == Relva && length obs == l                  = boaRelva ((ter,obs):t) x
      | proximosObstaculosValidos l ((ter, obs):t) == [] = obs
      | even (head(tail (randomizer x)))                 = (proximoObstaculo ((x*3)+1) l ((ter, (obs ++ [proxobst1])):t))
      | otherwise                                        = (proximoObstaculo ((x*3)+1) l ((ter, (obs ++ [proxobst2])):t))
                                                            where proxobst1 = last (proximosObstaculosValidos l ((ter, obs):t))
                                                                  proxobst2 = head (proximosObstaculosValidos l ((ter, obs):t))

    :: Int                         recebe um Int dado inicialmente pelo Usuário que servirá para fins de "aleatoriedade"
    -> Int                         recebe um Int relacionado há Largura do Mapa
    -> [(Terreno, [Obstaculo])]    recebe uma lista de tuplos de pares Terreno | lista de Obstáculos para perceber com que Terreno está a tratar 
    -> [Obstaculo]                 devolve uma lista de "Obstáculos" certa

"proximoObstaculo"irá usar as funções auxiliares "'randomizer x'" para decidir quais os "Obstáculos" irá adicionar para a lista final de obstáculos, vindos estes da função auxiliar "proximosObstaculosValidos"

proximosObstaculosValidos :: Int -> [(Terreno, [Obstaculo])] -> [Obstaculo]

proximosObstaculosValidos

A função "proximosObstaculosValidos" é uma função que, dado um Inteiro relacionado à Largura do Mapa, e um par Terreno | lista de Obstaculos, deduz uma lista de possíveis obstáculos com atenção às condições da primeira tarefa

Podemos assim definir "proximosObstaculosValidos" da seguinte maneira:

Definição dada

proximosObstaculosValidos :: Int -> [(Terreno, [Obstaculo])] -> [Obstaculo]
proximosObstaculosValidos larg ((_, obs):t) | larg == length obs = []
proximosObstaculosValidos larg ((Relva, obs):((Rio vel),obs2):t)
      | (vel && (not (elem Nenhum (take ((round((fromIntegral larg)/2))) obs)))) || (vel 0 && length obs < round((fromIntegral larg)/2) &&  (not (elem Nenhum obs)))
            = [Nenhum]
      | otherwise                                                                                                                            
            = [Nenhum, Arvore]

proximosObstaculosValidos larg ((Relva, obs):((Estrada vel),obs2):t)
      | (vel && (not (elem Nenhum (take ((round((fromIntegral larg)/2))) obs)))) || (vel 0 && length obs < round((fromIntegral larg)/2) &&  (not (elem Nenhum obs)))
            = [Nenhum]
      | otherwise                                                                                                                            
            = [Nenhum, Arvore]
proximosObstaculosValidos larg ((Relva, obs):t) = [Nenhum, Arvore]

proximosObstaculosValidos larg ((Rio vel, obs):t)
      |length obs == larg 
            = []
      |((drop (length obs - 2) obs) == [Nenhum,Tronco]) && (not (tronquinho obs 0)) 
      || (obs == [Tronco])
      || (obs /= []) && (not (elem Tronco (tail obs)) && ((length obs) >= larg-3)) 
            = [Tronco]
      |(tronquinho obs 0)
      || (((length obs) > larg-2) && (head obs == Tronco) && (tronquinho obs 0)) 
      || (((length obs) > larg-2) && (head obs == Nenhum)) 
            = [Nenhum]      
      |otherwise 
            = [Nenhum,Tronco]
            where tronquinho obs k |(obs == []) || (head obs == Nenhum && last obs == Nenhum) = k> 4
                                   |head obs == Tronco                                        = tronquinho (tail obs) (k+1)
                                   |last obs == Tronco                                        = tronquinho (init obs) (k+1)
                                   |otherwise                                                 = False

proximosObstaculosValidos larg ((Estrada vel, obs):t) 
      | (restaUmObs && (obs !! 1 == Carro)) || (restaDoisObs && head obs == Carro) || drop ((length obs) -3) obs == [Carro,Carro,Carro] || (drop ((length obs) -2) obs == [Carro,Nenhum])
            = [Nenhum]
      | otherwise                                                                                                                                
            = [Nenhum,Carro]
            where restaDoisObs = (larg - (length obs) <= 2)  
                  restaUmObs   = (larg - (length obs) <= 1)    

    :: Int                        recebe um Int relacionado à Largura
    -> [(Terreno, [Obstaculo])]   recebe uma lista de tuplos de pares Terreno | lista de Obstáculos para perceber com que Terreno está a tratar
                                      para perceber que "Obstáculos" pode devolver, tendo em conta as condições da 1ª tarefa 
    -> [Obstaculo]                devolve uma lista de "Obstáculos" possíveis

Baseando-se nas condições propostas na 1ª tarefa do projeto, a função "proximosObstaculosValidos" irá recursivamente ver todos os "Obstáculos" da lista e irá determinar uma lista de "Obstáculos" adequada ao "Terreno" dado, auxiliada pela função "proximosObstaculosValidos" concluir quais os Terrenos possíveis para continuar o Mapa (tendo em especial atenção a 1ª,3ª,4ª,5ª e 6ª condição da tarefa 1)

bomRio :: [Obstaculo] -> [Obstaculo]

bomRio

A função "bomRio" é uma função que, recebendo uma lista de Obstaculos, verifica recursivamente se neste existem Troncos, e caso isto seja verdadeiro, remove os Troncos que não estão acompanhados por pelo menos outro Tronco, tendo em conta, que posteriormente, ao querer que a parte gráfica descreva um "deslocamento suave", ultrapassar um Rio pode ser mais difícil com apenas 1 Tronco

Podemos assim definir "bomRio" da seguinte maneira:

Definição dada

bomRio :: [Obstaculo] -> [Obstaculo]
bomRio obs |length (elemIndices Nenhum obs) < 2 = obs
           |(head (tail (elemIndices Nenhum obs)))-(head (elemIndices Nenhum obs)) == 2 = (take (1+ head (elemIndices Nenhum obs)) obs) ++ [Nenhum] ++ (bomRio (drop (2+ head (elemIndices Nenhum obs)) obs)) 
           | otherwise = (take (1+ head (elemIndices Nenhum obs)) obs) ++ (bomRio (drop (1+ head (elemIndices Nenhum obs)) obs))

    :: [Obstaculo]   recebe uma lista de Obstaculos
    -> [Obstaculo]   devolve uma lista de Obstaculos

boaRelva :: [(Terreno, [Obstaculo])] -> Int -> [Obstaculo]

boaRelva

A função "boaRelva" ,tal como a função bomRio, verifica se a lista de Relvas recebida tem pelo menos "um caminho passável", caso contrário, chama a função actuallyBoaRelva e reescreve a lista de tuplos pares Terreno |lista de Obstaculos

Podemos assim definir "boaRelva" da seguinte maneira:

Definição dada

boaRelva :: [(Terreno,[Obstaculo])] -> Int -> [Obstaculo]
boaRelva ((Relva,obs):t) x | (elem (replicate ((length (numRelvas ((Relva, obs):t)))) Nenhum) (obsRelvas(numRelvas ((Relva, obs):t)))) = obs
                           | otherwise = actuallyBoaRelva (relvLinhas (numRelvas ((Relva,obs):t))) x 

    :: [(Terreno,[Obstaculo])]   recebe uma lista de tuplos pares Terreno | lista de Obstaculos
    -> Int recebe um Inteiro usado para aleatoriedade
    -> [Obstaculo]   devolve uma lista de Obstaculos

actuallyBoaRelva :: [[Obstaculo]] -> Int -> [Obstaculo]

posneg :: Int -> Int

posneg

A função "posneg" é uma função que, baseando-se na randomizer, decide se resulta num 1 ou num (-1), se o randomizer obtido pelo input do Usuário for par ou ímpar

Podemos assim definir "posneg" da seguinte maneira:

Definição dada

posneg :: Int -> Int
posneg x | odd (last (randomizer x)) = (-1)
         | otherwise = 1

    :: Int   recebe um Int relacionado à "aleatoriedade" da criação de uma nova linha 
    -> Int   devolve um Int (1 ou (-1))

randomizer :: Int -> [Int]

randomizer

A função "randomizer" é uma função que, dado um Int pelo Usuário, devolve um número "pseudo-aleatório"

Podemos assim definir "randomizer" da seguinte maneira:

Definição dada

randomizer :: Int -> [Int]
randomizer x 
      = take 5 $ randoms (mkStdGen x)

    :: Int     recebe um Int relacionado à "aleatoriedade" da criação de uma nova linha 
    -> [Int]   devolve uma lista de 2 números "pseudo-aleatórios"

numRelvas :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

numRelvas

A função "numRelvas" é uma função que, dado um tuplo de pares Terreno | lista de Obstaculos Devolve apenas a lista contida na original, que apenas contém Relvas

Definição dada

numRelvas :: [(Terreno,[Obstaculo])] -> [(Terreno,[Obstaculo])]
numRelvas ((Relva,obs):t) 
      = (Relva,obs):(numRelvas t)
numRelvas _ 
      = []

    :: [(Terreno,[Obstaculo])]     recebe uma lista de tuplos de pares Terreno | lista de Obstaculos
    -> [(Terreno,[Obstaculo])]   devolve uma lista de tuplos de pares Terreno | lista de Obstaculos

obsRelvas :: [(Terreno, [Obstaculo])] -> [[Obstaculo]]

obsRelvas

A função "obsRelvas" é uma função que, se considerarmos a lista de Obstaculos numa matriz, torna esta mesma matriz numa matriz transposta, com auxílio das funções relvLinhas e relvCols

Definição dada

obsRelvas :: [(Terreno,[Obstaculo])] -> [[Obstaculo]]
obsRelvas terobs 
      = relvCols (relvLinhas terobs)

    :: [(Terreno,[Obstaculo])]     recebe uma lista de tuplos de pares Terreno | lista de Obstaculos
    -> [[Obstaculo]]   devolve uma lista de listas de Obstaculos

relvLinhas :: [(Terreno, [Obstaculo])] -> [[Obstaculo]]

relvLinhas

A função "relvLinhas" é uma função que, dada uma lista de tuplos de pares Terreno | lista de Obstaculos, devolve uma lista de listas de Obstaculos

Definição dada

relvLinhas :: [(Terreno,[Obstaculo])] -> [[Obstaculo]]
relvLinhas []    
      = []
relvLinhas terobs 
      = (map snd terobs)

    :: [(Terreno,[Obstaculo])]     recebe um lista de tuplos de pares Terreno | lista de Obstaculos
    -> [[Obstaculo]]   devolve uma lista de listas de Obstaculos

relvCols :: [[Obstaculo]] -> [[Obstaculo]]

relvCols

A função "relvCols" é uma função que, dada uma lista de listas de Obstaculos, considerando-a como uma matriz, devolve a matriz transposta da mesma

Definição dada

relvCols :: [[Obstaculo]] -> [[Obstaculo]]
relvCols [] 
      = []
relvCols ([]:t) 
      = relvCols t
relvCols obsObs 
      = (map head obsObs):(relvCols (map tail obsObs))

    :: [[Obstaculo]]   recebe uma lista de listas de Obstaculos
    -> [[Obstaculo]]   devolve uma lista de listas de Obstaculos

velTer1 :: Terreno -> Int

velTer1

A função "velTer1" é uma função que, dado um Terreno, devolve a Velocidade deste

Definição dada

velTer1 :: Terreno -> Int
velTer1 (Estrada vel) = vel
velTer1 (Rio vel) = vel
velTer1 (Relva) = 0

    :: Terreno   recebe um Terreno
    -> Int   devolve um Int