Tarefa4_2022li1g124

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

jogoTerminou

A função "jogoTerminou" é uma função que, auxiliada pelas "atropelado", "foradelimites", "afogou", consegue receber um __Jogo__ e determinar se o Jogador ultrapassou um limite, afundou-se no rio ou se foi atropelado. Isto dado um "Mapa" e devolvendo um Booleano.

Podemos assim definir "jogoTerminou" da seguinte maneira:

Definição dada

jogoTerminou :: Jogo -> Bool
jogoTerminou (Jogo (Jogador (x,y)) (Mapa larg ((ter , obs):t))) = (foradelimites larg ((ter, obs):t) (x,y)) || (atropelado ((ter, obs):t) (x,y)) ||  (afogou ((ter, obs):t) (x,y))

foradelimites :: Int -> [(Terreno, [Obstaculo])] -> Coordenadas -> Bool

foradelimites

A função "foradelimites" é uma função que dado um "Int", um "[(Terreno, [Obstaculo])]" e "Coordenadas" determina se o Jogador encontra-se numa posição para lá dos limites do mapa, devolvendo um Booleano caso seja verdade.

Podemos assim definir "foradelimites" da seguinte maneira:

Definição dada

foradelimites :: Int -> [(Terreno, [Obstaculo])] -> Coordenadas -> Bool
foradelimites larg terobs (x,y) | (x || (x= larg) || (y || (y= (length terobs)) = True
                                | otherwise                                                   = False 

atropelado :: [(Terreno, [Obstaculo])] -> Coordenadas -> Bool

atropelado

A função "atropelado" é uma função que dado um "[(Terreno, [Obstaculo])]", determina se o Jogador encontra-se na mesma posição que um "Carro" e devolvendo um Booleano caso seja verdade.

Podemos assim definir "atropelado" da seguinte maneira:

Definição dada

atropelado :: [(Terreno, [Obstaculo])] -> Coordenadas -> Bool
atropelado ((ter, obs):t) (x,y) = ((snd (((ter , obs):t) !! y)) !! x) == Carro 

afogou :: [(Terreno, [Obstaculo])] -> Coordenadas -> Bool

afogou

A função "afogou" é uma função que dado um "[(Terreno, [Obstaculo])]" e "Coordenadas" determina se o Jogador encontra-se num Nenhum dentro de um "Rio", afogando-se assim, devolvendo um Booleano caso seja verdade. Esta apela pela função "serario" como auxiliar para determinar se se encontra num "Rio"

Podemos assim definir "afogou" da seguinte maneira:

Definição dada

afogou :: [(Terreno, [Obstaculo])] -> Coordenadas -> Bool
afogou terobs (x,y) = serario (fst (terobs !! y)) && (((snd (terobs !! y)) !! x) == Nenhum)

serario :: Terreno -> Bool

serario

A função "serario" é uma função que dado um "Terreno" e determina se o Jogador encontra-se num "Rio", devolvendo um Booleano caso seja verdade.

Podemos assim definir "serario" da seguinte maneira:

Definição dada

serario :: Terreno -> Bool
serario (Rio _) = True
serario (_)     = False